pyproject.toml
[project]
name = "translate-cloud-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-translate>=3.24.0",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]
